refactor(tools): split tools.py into tools/ package (mixins, ≤450 lines/file)
OdooToolHandler is now composed from per-domain mixins (formatting,
query, introspection, crud, bulk, binary, messaging); shared constants
and the _current_sub contextvar live in tools/_common.py. Registration
order is unchanged. tools/__init__.py re-exports the public surface
(OdooToolHandler, register_tools, _current_sub, MAX_BULK_SIZE) so all
existing import paths keep working.
