+++ mcp_server_odoo/tools/formatting.py
@@ -1,0 +1,348 @@
+"""Formatting and smart field selection helpers for Odoo tool handlers."""
+
+from __future__ import annotations
+
+from datetime import datetime
+from typing import Any, Dict, List, Optional
+
+from ..connection_protocol import OdooConnectionProtocol
+from ._common import logger
+
+
+class FormattingMixin:
+    """Datetime formatting and smart default field selection."""
+
+    def _format_datetime(self, value: str) -> str:
+        """Format datetime values to ISO 8601 with timezone."""
+        if not value or not isinstance(value, str):
+            return value
+
+        # Handle Odoo's compact datetime format (YYYYMMDDTHH:MM:SS)
+        if len(value) == 17 and "T" in value and "-" not in value:
+            try:
+                dt = datetime.strptime(value, "%Y%m%dT%H:%M:%S")
+                return dt.strftime("%Y-%m-%dT%H:%M:%S+00:00")
+            except ValueError:
+                pass
+
+        # Handle standard Odoo datetime format (YYYY-MM-DD HH:MM:SS)
+        if " " in value and len(value) == 19:
+            try:
+                dt = datetime.strptime(value, "%Y-%m-%d %H:%M:%S")
+                return dt.strftime("%Y-%m-%dT%H:%M:%S+00:00")
+            except ValueError:
+                pass
+
+        return value
+
+    def _process_record_dates(
+        self,
+        record: Dict[str, Any],
+        model: str,
+        connection: Optional[OdooConnectionProtocol] = None,
+    ) -> Dict[str, Any]:
+        """Process datetime fields in a record to ensure proper formatting."""
+        conn = connection or self.connection
+        # Common datetime field names in Odoo
+        known_datetime_fields = {
+            "create_date",
+            "write_date",
+            "date",
+            "datetime",
+            "date_start",
+            "date_end",
+            "date_from",
+            "date_to",
+            "date_order",
+            "date_invoice",
+            "date_due",
+            "last_update",
+            "last_activity",
+            "activity_date_deadline",
+        }
+
+        # First try to get field metadata
+        fields_info = None
+        try:
+            fields_info = conn.fields_get(model)
+        except Exception:
+            # Field metadata unavailable, will use fallback detection
+            pass
+
+        # Process each field in the record
+        for field_name, field_value in record.items():
+            if not isinstance(field_value, str):
+                continue
+
+            should_format = False
+
+            # Check if field is identified as datetime from metadata
+            if fields_info and isinstance(fields_info, dict) and field_name in fields_info:
+                field_type = fields_info[field_name].get("type")
+                if field_type == "datetime":
+                    should_format = True
+
+            # Check if field name suggests it's a datetime field
+            if not should_format and field_name in known_datetime_fields:
+                should_format = True
+
+            # Check if field name ends with common datetime suffixes
+            if not should_format and any(
+                field_name.endswith(suffix) for suffix in ["_date", "_datetime", "_time"]
+            ):
+                should_format = True
+
+            # Pattern-based detection for datetime-like strings
+            if not should_format and (
+                (
+                    len(field_value) == 17 and "T" in field_value and "-" not in field_value
+                )  # 20250607T21:55:52
+                or (
+                    len(field_value) == 19 and " " in field_value and field_value.count("-") == 2
+                )  # 2025-06-07 21:55:52
+            ):
+                should_format = True
+
+            # Apply formatting if needed
+            if should_format:
+                formatted = self._format_datetime(field_value)
+                if formatted != field_value:
+                    record[field_name] = formatted
+
+        return record
+
+    def _should_include_field_by_default(self, field_name: str, field_info: Dict[str, Any]) -> bool:
+        """Determine if a field should be included in default response.
+
+        Args:
+            field_name: Name of the field
+            field_info: Field metadata from fields_get()
+
+        Returns:
+            True if field should be included in default response
+        """
+        # Always include essential fields
+        always_include = {"id", "name", "display_name", "active", "company_id"}
+        if field_name in always_include:
+            return True
+
+        # Exclude system/technical fields by prefix
+        exclude_prefixes = ("_", "message_", "activity_", "website_message_")
+        if field_name.startswith(exclude_prefixes):
+            return False
+
+        # Exclude specific technical fields
+        exclude_fields = {
+            "write_date",
+            "create_date",
+            "write_uid",
+            "create_uid",
+            "__last_update",
+            "access_token",
+            "access_warning",
+            "access_url",
+        }
+        if field_name in exclude_fields:
+            return False
+
+        # Get field type
+        field_type = field_info.get("type", "")
+
+        # Exclude binary and large fields
+        if field_type in ("binary", "image", "html"):
+            return False
+
+        # Exclude expensive computed fields (non-stored)
+        if field_info.get("compute") and not field_info.get("store", True):
+            return False
+
+        # Exclude one2many and many2many fields (can be large)
+        if field_type in ("one2many", "many2many"):
+            return False
+
+        # Include required fields
+        if field_info.get("required"):
+            return True
+
+        # Include simple stored fields that are searchable
+        if field_info.get("store", True) and field_info.get("searchable", True):
+            if field_type in (
+                "char",
+                "text",
+                "boolean",
+                "integer",
+                "float",
+                "date",
+                "datetime",
+                "selection",
+                "many2one",
+            ):
+                return True
+
+        return False
+
+    def _score_field_importance(self, field_name: str, field_info: Dict[str, Any]) -> int:
+        """Score field importance for smart default selection.
+
+        Args:
+            field_name: Name of the field
+            field_info: Field metadata from fields_get()
+
+        Returns:
+            Importance score (higher = more important)
+        """
+        # Tier 1: Essential fields (always included)
+        if field_name in {"id", "name", "display_name", "active"}:
+            return 1000
+
+        # Exclude system/technical fields by prefix
+        exclude_prefixes = ("_", "message_", "activity_", "website_message_")
+        if field_name.startswith(exclude_prefixes):
+            return 0
+
+        # Exclude specific technical fields
+        exclude_fields = {
+            "write_date",
+            "create_date",
+            "write_uid",
+            "create_uid",
+            "__last_update",
+            "access_token",
+            "access_warning",
+            "access_url",
+        }
+        if field_name in exclude_fields:
+            return 0
+
+        score = 0
+
+        # Tier 2: Required fields are very important
+        if field_info.get("required"):
+            score += 500
+
+        # Tier 3: Field type importance
+        field_type = field_info.get("type", "")
+        type_scores = {
+            "char": 200,
+            "boolean": 180,
+            "selection": 170,
+            "integer": 160,
+            "float": 160,
+            "monetary": 140,
+            "date": 150,
+            "datetime": 150,
+            "many2one": 120,  # Relations useful but not primary
+            "text": 80,
+            "one2many": 40,
+            "many2many": 40,  # Heavy relations
+            "binary": 10,
+            "html": 10,
+            "image": 10,  # Heavy content
+        }
+        score += type_scores.get(field_type, 50)
+
+        # Tier 4: Storage and searchability bonuses
+        if field_info.get("store", True):
+            score += 80
+        if field_info.get("searchable", True):
+            score += 40
+
+        # Tier 5: Business-relevant field patterns (bonus)
+        business_patterns = [
+            "state",
+            "status",
+            "stage",
+            "priority",
+            "company",
+            "currency",
+            "amount",
+            "total",
+            "date",
+            "user",
+            "partner",
+            "email",
+            "phone",
+            "address",
+            "street",
+            "city",
+            "country",
+            "code",
+            "ref",
+            "number",
+        ]
+        if any(pattern in field_name.lower() for pattern in business_patterns):
+            score += 60
+
+        # Exclude expensive computed fields (non-stored)
+        if field_info.get("compute") and not field_info.get("store", True):
+            score = min(score, 30)  # Cap computed fields at low score
+
+        # Exclude large field types completely
+        if field_type in ("binary", "image", "html"):
+            return 0
+
+        # Exclude one2many and many2many fields (can be large)
+        if field_type in ("one2many", "many2many"):
+            return 0
+
+        return max(score, 0)
+
+    def _get_smart_default_fields(
+        self, model: str, connection: Optional[OdooConnectionProtocol] = None
+    ) -> Optional[List[str]]:
+        """Get smart default fields for a model using field importance scoring.
+
+        Args:
+            model: The Odoo model name
+            connection: Odoo connection to use (falls back to self.connection)
+
+        Returns:
+            List of field names to include by default, or None if unable to determine
+        """
+        conn = connection or self.connection
+        try:
+            # Get all field definitions
+            fields_info = conn.fields_get(model)
+
+            # Score all fields by importance
+            field_scores = []
+            for field_name, field_info in fields_info.items():
+                score = self._score_field_importance(field_name, field_info)
+                if score > 0:  # Only include fields with positive scores
+                    field_scores.append((field_name, score))
+
+            # Sort by score (highest first)
+            field_scores.sort(key=lambda x: x[1], reverse=True)
+
+            # Select top N fields based on configuration
+            max_fields = self.config.max_smart_fields
+            selected_fields = [field_name for field_name, _ in field_scores[:max_fields]]
+
+            # Ensure essential fields are always included
+            essential_fields = ["id", "name", "display_name", "active"]
+            for field in essential_fields:
+                if field in fields_info and field not in selected_fields:
+                    selected_fields.append(field)
+
+            # Remove duplicates while preserving order
+            final_fields = []
+            seen = set()
+            for field in selected_fields:
+                if field not in seen:
+                    final_fields.append(field)
+                    seen.add(field)
+
+            # Ensure we have at least essential fields
+            if not final_fields:
+                final_fields = [f for f in essential_fields if f in fields_info]
+
+            logger.debug(
+                f"Smart default fields for {model}: {len(final_fields)} of {len(fields_info)} fields "
+                f"(max configured: {max_fields})"
+            )
+            return final_fields
+
+        except Exception as e:
+            logger.warning(f"Could not determine default fields for {model}: {e}")
+            # Return None to indicate we should get all fields
+            return None
